+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * CLI: `find` command to look up game titles from the manifest.
     This incorporates the `--by-steam-id` option from the `backup`/etc commands
     and adds some new ones, like `--normalized` to look up games by an inexact name.
+  * CLI: Backup options: `--format`, `--compression`, `--full-limit`, `--differential-limit`.
   * On startup, Ludusavi will prune any useless blank configurations
     (e.g., roots with a blank path).
 * Changed: